Connect monitoring spans to runs, fixes
