+++ weave/op_def.py
@@ -30,6 +30,9 @@ from .language_features.tagging import (
 )
 from . import language_autocall
 
+if typing.TYPE_CHECKING:
+    from .run_streamtable_span import RunStreamTableSpan
+
 
 _no_refine: contextvars.ContextVar[bool] = contextvars.ContextVar(
     "_no_refine", default=False
@@ -603,7 +606,8 @@ class OpDef:
     def op_def_is_auto_tag_handling_arrow_op(self) -> bool:
         return isinstance(self, AutoTagHandlingArrowOpDef)
 
-    def runs(self) -> eager.WeaveIter[run.Run]:
+    # TODO: Should be generic Run protocol, but need to update graph_client type first.
+    def runs(self) -> eager.WeaveIter["RunStreamTableSpan"]:
         client = graph_client_context.require_graph_client()
         return client.op_runs(self)
