+++ weave/ops_arrow/list_ops.py
@@ -28,6 +28,7 @@ from . import arrow_tags
 from .vectorize import _apply_fn_node_with_tag_pushdown
 from . import convert
 from .convert import to_compare_safe
+from .concat import concatenate_all
 from .constructors import (
     vectorized_container_constructor_preprocessor,
     vectorized_input_types,
@@ -712,25 +713,7 @@ def concat(arr):
 
     # We merge the lists mergesort-style, which is `O(n*log(n))`
     # DO NOT merge the lists reduce-style, which is `O(n^2)`
-    return merge_concat(tagged)
-
-
-def merge_concat(arr: list[ArrowWeaveList]) -> ArrowWeaveList:
-    if len(arr) == 0:
-        raise ValueError("arr must not be empty")
-    if len(arr) == 1:
-        return arr[0]
-    left, right = merge_concat_split(arr)
-    return merge_concat(left).concat(merge_concat(right))
-
-
-def merge_concat_split(
-    arr: list[ArrowWeaveList],
-) -> tuple[list[ArrowWeaveList], list[ArrowWeaveList]]:
-    if len(arr) < 2:
-        raise ValueError("arr must have length of at least 2")
-    middle_index = len(arr) // 2
-    return arr[:middle_index], arr[middle_index:]
+    return concatenate_all(tagged)
 
 
 # # Putting this here instead of in number b/c it is just a map function
